LayoutTests/fast/images/inline-image-box-with-no-alt-should-collapse-no-quirks-expected.txt

STANDARDS mode.

1. 'alt' attribute is not set (image box should not collapse):
 20px 20px

2. 'alt' attribute is set to a non-empty value (image box should not collapse):
 20px 20px  20px 20px

3. 'alt' attribute is set to empty value (image box should collapse):
 0px 0px

4. 'alt' attribute is set to empty value on a blockified image (image box should not collapse):
20px 20px

5. 'alt' attribute is set to empty value and width/height values are set (image box should collapse):
width attribute is set: 0px 0px
height attribute is set: 0px 0px
width and height attributes are set: 0px 0px
inline style with width set: 0px 0px
inline style with height set: 0px 0px
inline style with width and height: 0px 0px
